+++ .github/workflows/pullrequest.yml
@@ -38,8 +38,8 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-maven-
 
-      - name: Build with Maven
-        run: mvn --batch-mode --update-snapshots verify -P integration-test
+      - name: Verify with Maven
+        run: mvn --batch-mode --update-snapshots verify -P e2e-test
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@1dd0ce34be62fac4f3b714f860c0b0c520acd35d
