+++ CONTRIBUTING.md
@@ -16,16 +16,16 @@ If you think we might be out of date with the spec, you can check that by invoki
 
 If you're adding tests to cover something in the spec, use the `@Specification` annotation like you see throughout the test suites.
 
-## Integration tests
+## End-to-End Tests
 
-The continuous integration runs a set of [gherkin integration tests](https://github.com/open-feature/test-harness/blob/main/features/evaluation.feature) using [`flagd`](https://github.com/open-feature/flagd). These tests do not run with the default maven profile. If you'd like to run them locally, you can start the flagd testbed with
+The continuous integration runs a set of [gherkin e2e tests](https://github.com/open-feature/test-harness/blob/main/features/evaluation.feature) using [`flagd`](https://github.com/open-feature/flagd). These tests do not run with the default maven profile. If you'd like to run them locally, you can start the flagd testbed with
 
 ```
 docker run -p 8013:8013 ghcr.io/open-feature/flagd-testbed:latest
 ```
 and then run 
 ```
-mvn test -P integration-test
+mvn test -P e2e-test
 ```
 
 ## Releasing
