+++ pom.xml
@@ -12,7 +12,7 @@
     <maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
     <junit.jupiter.version>5.9.3</junit.jupiter.version>
     <!-- exclusion expression for integration tests -->
-    <testExclusions>**/integration/*.java</testExclusions>
+    <testExclusions>**/e2e/*.java</testExclusions>
     <module-name>${groupId}.${artifactId}</module-name>
   </properties>
 
@@ -495,10 +495,10 @@
 
   <profiles>
     <profile>
-      <!-- this profile handles running the flagd integration tests -->
-      <id>integration-test</id>
+      <!-- this profile handles running the flagd e2e tests -->
+      <id>e2e-test</id>
       <properties>
-        <!-- run the integration tests by clearing the exclusions -->
+        <!-- run the e2e tests by clearing the exclusions -->
         <testExclusions/>
       </properties>  
       <build>
